+++ lmdoctor/extraction_utils.py
@@ -6,6 +6,8 @@ from .target_specific_utils.morality_utils import fetch_morality_data_conceptual
 from .target_specific_utils.emotion_utils import fetch_emotion_data_wrapper
 from .target_specific_utils.fairness_utils import fetch_fairness_data_conceptual_wrapper, fetch_fairness_data_functional_wrapper
 from .target_specific_utils.harmlessness_utils import fetch_harmlessness_data_conceptual
+from .probe_utils import probe_pca, probe_logreg, probe_massmean 
+
 
 import numpy as np
 from collections import defaultdict
@@ -19,10 +21,11 @@ logger = setup_logger()
 
 class Extractor:
     def __init__(self, model, tokenizer, user_tag, assistant_tag, device='cuda:0', 
-                 extraction_target=None, extraction_method=None, direction_op='pca',
+                 extraction_target=None, extraction_method=None, probe_type='pca',
                  **kwargs):
         """
         kwargs: Additional keyword arguments, such as:
+        - probe_type (str): The computation used to find directions. Supports pca, massmean, logreg
         - bias_type (str): The type of bias to use for fairness extraction.
         - emotion_type (str): The type of emotion to use for emotion extraction.
         - n_trim_tokens (int): The number of tokens to trim from the end of statements for functional extraction. Defaults to 5.
@@ -38,7 +41,7 @@ class Extractor:
         self.assistant_tag = assistant_tag
         self.extraction_target = extraction_target
         self.extraction_method = extraction_method
-        self.direction_op = direction_op
+        self.probe_type = probe_type
         self.device = device
         self.kwargs = kwargs
         self.direction_info = None
@@ -54,7 +57,7 @@ class Extractor:
             self.user_tag, self.assistant_tag, n_train_pairs, n_dev_pairs, n_test_pairs, **self.kwargs)
         self.train_acts = get_activations_for_paired_statements(
             self.statement_pairs['train'], self.model, self.tokenizer, batch_size, device=self.device)   
-        self.direction_info = get_directions(self.train_acts, self.device, self.direction_op)
+        self.direction_info = get_directions(self.train_acts, self.device, self.probe_type)
     
 
 def get_extraction_function(target, extraction_method=None, **kwargs):
@@ -195,117 +198,18 @@ def get_activations_for_paired_statements(statement_pairs, model, tokenizer, bat
     return layer_to_act_pairs  
 
 
-def scale_direction(act_pairs, direction):
-    # scale direction such that p(mu_b + scaled_direction) = p(mu_a), following marks et al
-    # (https://arxiv.org/abs/2310.06824)
-    mu_a = torch.mean(act_pairs[:, 0, :], axis=0)
-    mu_b = torch.mean(act_pairs[:, 1, :], axis=0)
-    norm_direction = F.normalize(direction, dim=0)
-    diff_proj = (mu_a - mu_b) @ norm_direction.view(-1)
-    scaled_direction = (norm_direction * diff_proj).view(-1)
-    return scaled_direction
+def get_directions(train_acts, device, probe_type):
+    if probe_type == 'pca':
+        return probe_pca(train_acts, device)
+    elif probe_type == 'logreg':
+        return probe_logreg(train_acts, device)
+    elif probe_type == 'massmean':
+        return probe_massmean(train_acts, device)
+    else:
+        probe_types = ['pca', 'logreg', 'massmean']
+        raise ValueError(f'probe_type must be one of {probe_types} but is {probe_type}.')
 
-
-def get_directions(train_acts, device, method):
-    if method == 'pca':
-        return get_directions_pca(train_acts, device)
-    elif method == 'logreg':
-        return get_directions_logreg(train_acts, device)
-    elif method == 'mass_mean':
-        return get_directions_massmean(train_acts, device)
         
-
-def get_directions_pca(train_acts, device):
-    directions = {}
-    scaled_directions = {}
-    mean_diffs = {}
-    direction_info = defaultdict(dict)
-    for layer in train_acts:
-        act_pairs = train_acts[layer]
-        shuffled_pairs = [] # must shuffle pairs to create variability in difference between pairs
-        for pair in act_pairs:
-            pair = pair[torch.randperm(2)]
-            shuffled_pairs.append(pair)
-        shuffled_pairs = torch.stack(shuffled_pairs)
-        diffs = shuffled_pairs[:, 0, :] - shuffled_pairs[:, 1, :] 
-        mean_diffs[layer] = torch.mean(diffs, axis=0)
-        centered_diffs = diffs - mean_diffs[layer] # is centering necessary?
-        pca = PCA(n_components=1)
-        pca.fit(centered_diffs.detach().cpu())
-        direction = torch.tensor(pca.components_[0], dtype=act_pairs.dtype).to(device)
-        directions[layer] = direction 
-        # scale
-        act_pairs = train_acts[layer]
-        scaled_direction = scale_direction(act_pairs, direction)
-        scaled_directions[layer] = scaled_direction
-    
-    direction_info['unscaled_directions'] = directions # these aren't used, but kept for posterity
-    direction_info['directions'] = scaled_directions
-    direction_info['mean_diffs'] = mean_diffs
-    return direction_info
-
-
-def get_directions_logreg(train_acts, device):
-    directions = {}
-    scaled_directions = {}
-    direction_info = defaultdict(dict)
-    for layer in train_acts:
-        act_pairs = train_acts[layer]
-        n_pairs, n_classes, hidden_dim = act_pairs.shape
-        acts = act_pairs.view(-1, hidden_dim)
-        labels = [1, 0] * n_pairs
-        clf = train_logreg(acts.cpu().numpy(), np.array(labels))
-        direction = torch.tensor(clf.coef_[0], dtype=act_pairs.dtype).to(device)
-        directions[layer] = direction
-        # scale
-        scaled_direction = scale_direction(act_pairs, direction)
-        scaled_directions[layer] = scaled_direction
-
-    direction_info['unscaled_directions'] = directions # these aren't used, but kept for posterity
-    direction_info['directions'] = scaled_directions
-    return direction_info
-    
-
-def train_logreg(X, y, test_clf=False):
-    from sklearn.model_selection import train_test_split
-    from sklearn.preprocessing import StandardScaler
-    from sklearn.linear_model import LogisticRegression
-    from sklearn.metrics import classification_report
-
-    if test_clf:
-        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)  # 70% training and 30% testing
-        clf = LogisticRegression(random_state=42)
-        clf.fit(X_train, y_train)
-        y_pred = clf.predict(X_test)    
-        print(classification_report(y_test, y_pred))
-        
-    # fit on all data
-    clf = LogisticRegression(random_state=42)
-    clf.fit(X, y)
-
-    return clf
-
-
-def get_directions_massmean(train_acts, device):
-    directions = {}
-    scaled_directions = {}
-    direction_info = defaultdict(dict)
-    for layer in train_acts:
-        act_pairs = train_acts[layer]
-        mu_a = torch.mean(act_pairs[:, 0, :], axis=0)
-        mu_b = torch.mean(act_pairs[:, 1, :], axis=0)
-        direction = mu_a - mu_b
-        directions[layer] = direction
-        # scale
-        scaled_direction = scale_direction(act_pairs, direction)
-        scaled_directions[layer] = scaled_direction
-
-    direction_info['unscaled_directions'] = directions # these aren't used, but kept for posterity
-    direction_info['directions'] = scaled_directions
-    return direction_info
-    
-    
-
 # def get_accs_for_pairs(test_acts, direction_info):
 #     """
 #     This could be used for testing accuracy of on the extraction dataset (in-distribution).
